model updated

+++ app/vector_store.py
@@ -4,7 +4,7 @@ from sentence_transformers import SentenceTransformer
 
 class VectorStore:
     def __init__(self):
-        self.model = SentenceTransformer("all-MiniLM-L6-v2")
+        self.model = SentenceTransformer("all-MiniLM-L6-v2")  # Small, fast
         self.texts = []
         self.meta = []
         self.index = faiss.IndexFlatL2(384)
@@ -13,12 +13,12 @@ class VectorStore:
         for doc in docs:
             chunks = [doc["text"][i:i+500] for i in range(0, len(doc["text"]), 500)]
             embeddings = self.model.encode(chunks)
-            self.index.add(np.array(embeddings))
+            self.index.add(np.array(embeddings).astype("float32"))
             self.texts.extend(chunks)
             self.meta.extend([{"source": doc["filename"], "chunk_id": i} for i in range(len(chunks))])
 
     def query(self, question, top_k=3):
-        q_embed = self.model.encode([question])
+        q_embed = self.model.encode([question]).astype("float32")
         D, I = self.index.search(np.array(q_embed), top_k)
         results = []
         for i in I[0]:
